рассовал компоненты и стили по разным папкам, удалил файлы заглушки, которые точно не нужны
